+++ reptar/calculators/drivers.py
@@ -37,9 +37,6 @@ class DriverEnergy:
 
     def __init__(
         self,
-        Z,
-        R,
-        E,
         worker,
         worker_kwargs,
         use_ray=False,
@@ -53,18 +50,6 @@ class DriverEnergy:
         r"""
         Parameters
         ----------
-        Z : :obj:`numpy.ndarray`, ndim: ``1``
-            Atomic numbers of the atoms with respect to ``R``. Should have shape
-            ``(i,)`` where ``i`` is the number of atoms in the system.
-        R : :obj:`numpy.ndarray`, ndim: ``3``
-            Cartesian coordinates of all structures in group. Should have shape
-            ``(j, i, 3)`` where ``j`` is the number of structures. Units are in
-            Angstroms.
-        E : :obj:`numpy.ndarray`, ndim: ``1``
-            Total electronic energies of all structures in ``R``. Energies that need
-            to be calculated should have ``NaN`` in the element corresponding
-            to the same index in ``R``. Should have shape ``(j,)``. Units are in
-            Hartrees.
         worker : ``callable``
             The desired worker function to compute energy and gradients. It should
             be the same as any previous calculations. The ``ray.remote`` decorator
@@ -91,16 +76,6 @@ class DriverEnergy:
         ray_address : :obj:`str`, default: ``'auto'``
             Ray cluster address to connect to.
         """
-        # Check arrays (obsessively)
-        assert R.ndim == 3
-        assert Z.shape[0] == R.shape[1]
-        assert R.shape[0] == E.shape[0]
-        assert R.shape[2] == 3
-
-        # Storing arrays and other information
-        self.Z = Z
-        self.R = R
-        self.E = E
         self.worker = worker
         self.worker_kwargs = worker_kwargs
         self.start_slice = start_slice
@@ -115,10 +90,8 @@ class DriverEnergy:
             if not ray.is_initialized():
                 ray.init(address=ray_address)
 
-            self.Z = ray.put(Z)
-            self.R = ray.put(R)
-
-    def _idx_todo(self):
+    @staticmethod
+    def _idx_todo(E, start_slice=None, end_slice=None):
         r"""Indices of missing energies (calculations to do).
 
         Determines this by finding all ``NaN`` elements.
@@ -128,37 +101,61 @@ class DriverEnergy:
         :obj:`numpy.ndarray`
             Indices for ``R`` that are missing energie values.
         """
-        E = self.E[self.start_slice : self.end_slice]
+        # pylint: disable-next=invalid-name
+        E = E[start_slice:end_slice]
         idx_todo = np.argwhere(np.isnan(E))[:, 0]
         return idx_todo
 
-    def run(self, saver=None):
+    def run(self, Z, R, E, saver=None):
         r"""Run the calculations.
 
         Parameters
         ----------
+        Z : :obj:`numpy.ndarray`, ndim: ``1``
+            Atomic numbers of the atoms with respect to ``R``. Should have shape
+            ``(i,)`` where ``i`` is the number of atoms in the system.
+        R : :obj:`numpy.ndarray`, ndim: ``3``
+            Cartesian coordinates of all structures in group. Should have shape
+            ``(j, i, 3)`` where ``j`` is the number of structures. Units are in
+            Angstroms.
+        E : :obj:`numpy.ndarray`, ndim: ``1``
+            Total electronic energies of all structures in ``R``. Energies that need
+            to be calculated should have ``NaN`` in the element corresponding
+            to the same index in ``R``. Should have shape ``(j,)``. Units are in
+            Hartrees.
         saver : :obj:`reptar.Saver`, optional
-            Save data after every worker finishes.
+            Save data after every worker finishes. Passes ``E`` into
+            :meth:`~reptar.Saver.save`.
 
         Returns
         -------
         :obj:`numpy.ndarray`
             The total electronic energy array, ``E``, after all computations.
         """
+        # Check arrays (obsessively)
+        assert R.ndim == 3
+        assert Z.shape[0] == R.shape[1]
+        assert R.shape[0] == E.shape[0]
+        assert R.shape[2] == 3
+
         worker = self.worker
-        idxs_todo = self._idx_todo()
+        idxs_todo = self._idx_todo(E, self.start_slice, self.end_slice)
         chunker = chunk_iterable(idxs_todo, self.chunk_size)
 
         if not self.use_ray:
             for idx in idxs_todo:
                 # pylint: disable-next=invalid-name
-                _, E_done = worker([idx], self.Z, self.R, **self.worker_kwargs)
-                self.E[idx] = E_done
+                _, E_done = worker([idx], Z, R, **self.worker_kwargs)
+                print(E_done)
+                E[idx] = E_done
 
                 if saver is not None:
-                    saver.save((self.E))
+                    saver.save(E)
         else:
             worker = ray.remote(worker)
+            Z = ray.put(Z)
+            R = ray.put(R)
+
             # Initialize ray workers
             workers = []
 
@@ -167,7 +164,7 @@ class DriverEnergy:
                     chunk = list(next(chunker))
                     workers.append(
                         worker.options(num_cpus=self.n_cpus_per_worker).remote(
-                            chunk, self.Z, self.R, **self.worker_kwargs
+                            chunk, Z, R, **self.worker_kwargs
                         )
                     )
                 except StopIteration:
@@ -181,14 +178,18 @@ class DriverEnergy:
                 done_id, workers = ray.wait(workers)
 
                 idx_done, E_done = ray.get(done_id)[0]  # pylint: disable=invalid-name
-                self.E[idx_done] = E_done
+                E[idx_done] = E_done
 
                 if saver is not None:
-                    saver.save((self.E))
+                    saver.save(E)
 
                 add_worker(workers, chunker)
 
-        return self.E
+            # Cleanup ray object store
+            del Z
+            del R
+
+        return E
 
 
 class DriverEnGrad:
@@ -199,10 +200,6 @@ class DriverEnGrad:
 
     def __init__(
         self,
-        Z,
-        R,
-        E,
-        G,
         worker,
         worker_kwargs,
         use_ray=False,
@@ -214,6 +211,68 @@ class DriverEnGrad:
         ray_address="auto",
     ):
         r"""
+        Parameters
+        ----------
+        worker : ``callable``
+            The desired worker function to compute energy and gradients. It should
+            be the same as any previous calculations. The ``ray.remote`` decorator
+            will be applied automatically.
+        worker_kwargs : :obj:`tuple`, ndim: ``1``
+            The other keyword arguments for the worker function after
+            ``R_idxs``, ``Z``, and ``R``.
+        use_ray : :obj:`bool`, default: ``False``
+            Use ray to parallelize calculations. If ``False``, calculations are
+            done serially. ``False`` can be useful when running locally or only
+            a few calculations are needed. ``True`` is useful for tons of
+            calculations.
+        n_workers : :obj:`int`, default: ``1``
+            Number of parallel workers to use if ``use_ray`` is ``True``.
+        n_cpus_per_worker : :obj:`int`, default: ``1``
+            Number of CPU cores to provide each worker.
+        chunk_size : :obj:`int`, default: ``50``
+            Number of calculations per task to do. This should be enough to make
+            the ray task overhead significantly less than calculations.
+        start_slice : :obj:`int`, default: ``None``
+            Trims ``R`` to start at this index.
+        end_slice : :obj:`int`, default: ``None``
+            Trims ``R`` to end at this index.
+        ray_address : :obj:`str`, default: ``'auto'``
+            Ray cluster address to connect to.
+        """
+        # Storing arrays and other information
+        self.worker = worker
+        self.worker_kwargs = worker_kwargs
+        self.start_slice = start_slice
+        self.end_slice = end_slice
+
+        self.n_workers = n_workers
+        self.n_cpus_per_worker = n_cpus_per_worker
+        self.chunk_size = chunk_size
+
+        self.use_ray = use_ray
+        if use_ray:
+            if not ray.is_initialized():
+                ray.init(address=ray_address)
+
+    @staticmethod
+    def _idx_todo(E, start_slice=None, end_slice=None):
+        r"""Indices of missing energies (calculations to do).
+
+        Determines this by finding all ``NaN`` elements.
+
+        Returns
+        -------
+        :obj:`numpy.ndarray`
+            Indices for ``R`` that are missing energie values.
+        """
+        # pylint: disable-next=invalid-name
+        E = E[start_slice:end_slice]
+        idx_todo = np.argwhere(np.isnan(E))[:, 0]
+        return idx_todo
+
+    def run(self, Z, R, E, G, saver=None):
+        r"""Run the calculations.
+
         Parameters
         ----------
         Z : :obj:`numpy.ndarray`, ndim: ``1``
@@ -232,31 +291,16 @@ class DriverEnGrad:
             Atomic gradients of all structures in ``R``. Gradients that need to be
             calculated should have ``NaN`` in the corresponding elements. Should
             have the same shape as ``R``. Units are in Hartrees/Angstrom.
-        worker : ``callable``
-            The desired worker function to compute energy and gradients. It should
-            be the same as any previous calculations. The ``ray.remote`` decorator
-            will be applied automatically.
-        worker_kwargs : :obj:`tuple`, ndim: ``1``
-            The other keyword arguments for the worker function after
-            ``R_idxs``, ``Z``, and ``R``.
-        use_ray : :obj:`bool`, default: ``False``
-            Use ray to parallelize calculations. If ``False``, calculations are
-            done serially. ``False`` can be useful when running locally or only
-            a few calculations are needed. ``True`` is useful for tons of
-            calculations.
-        n_workers : :obj:`int`, default: ``1``
-            Number of parallel workers to use if ``use_ray`` is ``True``.
-        n_cpus_per_worker : :obj:`int`, default: ``1``
-            Number of CPU cores to provide each worker.
-        chunk_size : :obj:`int`, default: ``50``
-            Number of calculations per task to do. This should be enough to make
-            the ray task overhead significantly less than calculations.
-        start_slice : :obj:`int`, default: ``None``
-            Trims ``R`` to start at this index.
-        end_slice : :obj:`int`, default: ``None``
-            Trims ``R`` to end at this index.
-        ray_address : :obj:`str`, default: ``'auto'``
-            Ray cluster address to connect to.
+        saver : :obj:`reptar.Saver`, optional
+            Save data after every worker finishes. Passes ``E, G`` into
+            :meth:`~reptar.Saver.save`.
+
+        Returns
+        -------
+        :obj:`numpy.ndarray`
+            The total electronic energy array, ``E``, after all computations.
+        :obj:`numpy.ndarray`
+            The atomic gradients array, ``E``, after all computations.
         """
         # Check arrays (obsessively)
         assert R.ndim == 3
@@ -267,73 +311,24 @@ class DriverEnGrad:
         assert R.shape[2] == 3
         assert G.shape[2] == 3
 
-        # Storing arrays and other information
-        self.Z = Z
-        self.R = R
-        self.E = E
-        self.G = G
-        self.worker = worker
-        self.worker_kwargs = worker_kwargs
-        self.start_slice = start_slice
-        self.end_slice = end_slice
-
-        self.n_workers = n_workers
-        self.n_cpus_per_worker = n_cpus_per_worker
-        self.chunk_size = chunk_size
-
-        self.use_ray = use_ray
-        if use_ray:
-            if not ray.is_initialized():
-                ray.init(address=ray_address)
-
-            self.Z = ray.put(Z)
-            self.R = ray.put(R)
-
-    def _idx_todo(self):
-        r"""Indices of missing energies (calculations to do).
-
-        Determines this by finding all ``NaN`` elements.
-
-        Returns
-        -------
-        :obj:`numpy.ndarray`
-            Indices for ``R`` that are missing energie values.
-        """
-        # pylint: disable-next=invalid-name
-        E = self.E[self.start_slice : self.end_slice]
-        idx_todo = np.argwhere(np.isnan(E))[:, 0]
-        return idx_todo
-
-    def run(self, saver=None):
-        r"""Run the calculations.
-
-        Parameters
-        ----------
-        saver : :obj:`reptar.Saver`, optional
-            Save data after every worker finishes.
-
-        Returns
-        -------
-        :obj:`numpy.ndarray`
-            The total electronic energy array, ``E``, after all computations.
-        :obj:`numpy.ndarray`
-            The atomic gradients array, ``E``, after all computations.
-        """
         worker = self.worker
-        idxs_todo = self._idx_todo()
+        idxs_todo = self._idx_todo(E, self.start_slice, self.end_slice)
         chunker = chunk_iterable(idxs_todo, self.chunk_size)
 
         if not self.use_ray:
             for idx in idxs_todo:
                 # pylint: disable-next=invalid-name
-                _, E_done, G_done = worker([idx], self.Z, self.R, **self.worker_kwargs)
-                self.E[idx] = E_done
-                self.G[idx] = G_done
+                _, E_done, G_done = worker([idx], Z, R, **self.worker_kwargs)
+                E[idx] = E_done
+                G[idx] = G_done
 
                 if saver is not None:
-                    saver.save((self.E, self.G))
+                    saver.save((E, G))
         else:
             worker = ray.remote(worker)
+            Z = ray.put(Z)
+            R = ray.put(R)
+
             # Initialize ray workers
             workers = []
 
@@ -342,7 +337,7 @@ class DriverEnGrad:
                     chunk = list(next(chunker))
                     workers.append(
                         worker.options(num_cpus=self.n_cpus_per_worker).remote(
-                            chunk, self.Z, self.R, **self.worker_kwargs
+                            chunk, Z, R, **self.worker_kwargs
                         )
                     )
                 except StopIteration:
@@ -357,15 +352,19 @@ class DriverEnGrad:
 
                 # pylint: disable-next=invalid-name
                 idx_done, E_done, G_done = ray.get(done_id)[0]
-                self.E[idx_done] = E_done
-                self.G[idx_done] = G_done
+                E[idx_done] = E_done
+                G[idx_done] = G_done
 
                 if saver is not None:
-                    saver.save((self.E, self.G))
+                    saver.save(E, G)
 
                 add_worker(workers, chunker)
 
-        return self.E, self.G
+            # Cleanup ray object store
+            del Z
+            del R
+
+        return E, G
 
 
 class DriverOpt:
@@ -376,11 +375,6 @@ class DriverOpt:
 
     def __init__(
         self,
-        Z,
-        R,
-        R_opt,  # pylint: disable=invalid-name
-        E,
-        G,
         worker,
         worker_kwargs,
         use_ray=False,
@@ -394,25 +388,6 @@ class DriverOpt:
         r"""
         Parameters
         ----------
-        Z : :obj:`numpy.ndarray`, ndim: ``1``
-            Atomic numbers of the atoms with respect to ``R``. Should have shape
-            ``(i,)`` where ``i`` is the number of atoms in the system.
-        R : :obj:`numpy.ndarray`, ndim: ``3``
-            Cartesian coordinates of all structures in group. Should have shape
-            ``(j, i, 3)`` where ``j`` is the number of structures. Units are in
-            Angstroms.
-        R_opt : :obj:`numpy.ndarray`, ndim: ``3``
-            Optimized Cartesian coordinates of all structures in group. This
-            is used to identifying which optimizations need to be done.
-        E : :obj:`numpy.ndarray`, ndim: ``1``
-            Total electronic energies of all structures in ``R``. Energies that need
-            to be calculated should have ``NaN`` in the element corresponding
-            to the same index in ``R``. Should have shape ``(j,)``. Units are in
-            Hartrees.
-        G : :obj:`numpy.ndarray`, ndim: ``3``
-            Atomic gradients of all structures in ``R``. Gradients that need to be
-            calculated should have ``NaN`` in the corresponding elements. Should
-            have the same shape as ``R``. Units are in Hartrees/Angstrom.
         worker : ``callable``
             The desired worker function to compute energy and gradients. It should
             be the same as any previous calculations. The ``ray.remote`` decorator
@@ -437,20 +412,10 @@ class DriverOpt:
         end_slice : :obj:`int`, default: ``None``
             Trims ``R`` to end at this index.
         ray_address : :obj:`str`, default: ``'auto'``
-            Ray cluster address to connect to.
+            Ray cluster address to connect to. Passes ``opt_conv, R_opt, E, G`` into
+            :meth:`~reptar.Saver.save`.
         """
-        # Check arrays (obsessively)
-        assert R.ndim == 3
-        assert Z.shape[0] == R.shape[1]
-        assert R.shape[2] == 3
-
         # Storing arrays and other information
-        self.Z = Z
-        self.R = R
-        self.R_opt = R_opt  # pylint: disable=invalid-name
-        self.opt_conv = ~np.isnan(self.R_opt[:, 0, 0])
-        self.E = E
-        self.G = G
         self.worker = worker
         self.worker_kwargs = worker_kwargs
         self.start_slice = start_slice
@@ -465,10 +430,8 @@ class DriverOpt:
             if not ray.is_initialized():
                 ray.init(address=ray_address)
 
-            self.Z = ray.put(Z)
-            self.R = ray.put(R)
-
-    def _idx_todo(self):
+    @staticmethod
+    def _idx_todo(opt_conv, start_slice=None, end_slice=None):
         r"""Indices of NaN geometries (calculations to do).
 
         Returns
@@ -476,13 +439,33 @@ class DriverOpt:
         :obj:`numpy.ndarray`
             Indices for ``R`` that are missing energie values.
         """
-        return np.where(~self.opt_conv[self.start_slice : self.end_slice])[0]
+        return np.where(~opt_conv[start_slice:end_slice])[0]
 
-    def run(self, saver=None):
+    # pylint: disable-next=invalid-name
+    def run(self, Z, R, R_opt, E, G, saver=None):
         r"""Run the calculations.
 
         Parameters
         ----------
+        Z : :obj:`numpy.ndarray`, ndim: ``1``
+            Atomic numbers of the atoms with respect to ``R``. Should have shape
+            ``(i,)`` where ``i`` is the number of atoms in the system.
+        R : :obj:`numpy.ndarray`, ndim: ``3``
+            Cartesian coordinates of all structures in group. Should have shape
+            ``(j, i, 3)`` where ``j`` is the number of structures. Units are in
+            Angstroms.
+        R_opt : :obj:`numpy.ndarray`, ndim: ``3``
+            Optimized Cartesian coordinates of all structures in group. This
+            is used to identifying which optimizations need to be done.
+        E : :obj:`numpy.ndarray`, ndim: ``1``
+            Total electronic energies of all structures in ``R``. Energies that need
+            to be calculated should have ``NaN`` in the element corresponding
+            to the same index in ``R``. Should have shape ``(j,)``. Units are in
+            Hartrees.
+        G : :obj:`numpy.ndarray`, ndim: ``3``
+            Atomic gradients of all structures in ``R``. Gradients that need to be
+            calculated should have ``NaN`` in the corresponding elements. Should
+            have the same shape as ``R``. Units are in Hartrees/Angstrom.
         saver : :obj:`reptar.Saver`, optional
             Save data after every worker finishes.
 
@@ -493,25 +476,35 @@ class DriverOpt:
         :obj:`numpy.ndarray`
             The atomic gradients array, ``E``, after all computations.
         """
+        # Check arrays (obsessively)
+        assert R.ndim == 3
+        assert Z.shape[0] == R.shape[1]
+        assert R.shape[2] == 3
+
+        opt_conv = ~np.isnan(R_opt[:, 0, 0])
+
         worker = self.worker
-        idxs_todo = self._idx_todo()
+        idxs_todo = self._idx_todo(opt_conv, self.start_slice, self.end_slice)
         chunker = chunk_iterable(idxs_todo, self.chunk_size)
 
         if not self.use_ray:
             for idx in idxs_todo:
                 # pylint: disable-next=invalid-name
                 _, opt_conv_done, R_opt_done, E_done, G_done = worker(
-                    [idx], self.Z, self.R, **self.worker_kwargs
+                    [idx], Z, R, **self.worker_kwargs
                 )
-                self.opt_conv[idx] = opt_conv_done
-                self.R_opt[idx] = R_opt_done
-                self.E[idx] = E_done
-                self.G[idx] = G_done
+                opt_conv[idx] = opt_conv_done
+                R_opt[idx] = R_opt_done
+                E[idx] = E_done
+                G[idx] = G_done
 
                 if saver is not None:
-                    saver.save((self.opt_conv, self.R_opt, self.E, self.G))
+                    saver.save(opt_conv, R_opt, E, G)
         else:
             worker = ray.remote(worker)
+            Z = ray.put(Z)
+            R = ray.put(R)
+
             # Initialize ray workers
             workers = []
 
@@ -520,7 +513,7 @@ class DriverOpt:
                     chunk = list(next(chunker))
                     workers.append(
                         worker.options(num_cpus=self.n_cpus_per_worker).remote(
-                            chunk, self.Z, self.R, **self.worker_kwargs
+                            chunk, Z, R, **self.worker_kwargs
                         )
                     )
                 except StopIteration:
@@ -537,14 +530,14 @@ class DriverOpt:
                 idxs_done, opt_conv_done, R_opt_done, E_done, G_done = ray.get(done_id)[
                     0
                 ]
-                self.opt_conv[idxs_done] = opt_conv_done
-                self.R_opt[idxs_done] = R_opt_done
-                self.E[idxs_done] = E_done
-                self.G[idxs_done] = G_done
+                opt_conv[idxs_done] = opt_conv_done
+                R_opt[idxs_done] = R_opt_done
+                E[idxs_done] = E_done
+                G[idxs_done] = G_done
 
                 if saver is not None:
-                    saver.save((self.opt_conv, self.R_opt, self.E, self.G))
+                    saver.save(opt_conv, R_opt, E, G)
 
                 add_worker(workers, chunker)
 
-        return self.opt_conv, self.R_opt, self.E, self.G
+        return opt_conv, R_opt, E, G