+++ tests/paths.py
@@ -22,7 +22,6 @@
 
 """Paths to files used for tests."""
 
-import pytest
 import os
 
 data_path = "../examples/reptar-data/"